move auth_noise into noise
